requirements.txt
feedparser==6.0.10
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
PyJWT==2.9.0
google-generativeai==0.7.2
APScheduler==3.10.4
mysql-connector-python==9.0.0
trafilatura==1.9.0
python-dotenv==1.0.1
requests==2.32.3
